XWIKI-14672: Refactoring jobs don't take into account wiki syntax in Macros
  * Fix package name
