+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-default/src/main/resources/META-INF/components.txt
@@ -8,5 +8,5 @@ org.xwiki.refactoring.internal.MainWikiRefactoringConfigurationSource
 org.xwiki.refactoring.internal.RefactoringConfigurationSource
 org.xwiki.refactoring.job.XClassDeletingListener
 org.xwiki.refactoring.internal.DefaultMacroRefactoring
-org.xwiki.refactoring.internal.util.DefaultReferenceRenamer
-org.xwiki.refactoring.internal.util.ResourceReferenceRenamer
+org.xwiki.refactoring.internal.DefaultReferenceRenamer
+org.xwiki.refactoring.internal.ResourceReferenceRenamer
